Moved supporting map utilies to common.tools.py
